inha win8 metrowin contribution / "porting mode" programming - many IDEs in different platforms and OSes... - m_build contribution...
Eu te amo Carlinhos, Carlos Gustavo Cecyn Lundgren!!

Obrigado Mummi!! Feliz 61 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 7 Ultimate 64-bit Service Pack 1 - with Portuguese - Brazil - Language for non-Unicode programs - ~10Mpbs NET Virtua Internet at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Mummi!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabi Cecyn!!
  Tack Rafael Cecyn Lundgren!!
  Tack Simon Petrus Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
}

Obrigado Cits
{
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Buhrer!!
}

Obrigado Brasil
{
   Obrigado guarda legal do CITS noturno 1!!
   Obrigado guarda legal do CITS noturno 2!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado piriguetes and vampeguetes...
   Obrigado operários eficientes...
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Dilma Souza!!
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}

Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt!!
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
T Obrigado Telminho, você é legal!!
1 Obrigado Telminho, você é bom!!
2 Obrigado Telminho, eu gosto de você!!
n Obrigado Telminho, eu gosto muito de você!!
n Obrigado Telminho, você é demais!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/c/metrowin/c_os_file_watcher.cpp
@@ -46,7 +46,7 @@ namespace file_watcher
       property ::Platform::String ^                                  m_strDirName;
       property size_t                                                m_pwatcher;   // should be exactly file_watcher_impl *
       property size_t                                                m_plistener;  // should be exactly file_watch_listener *
-      property size_t                                                m_id;
+      property uint64_t                                              m_id;
 
       static watch_struct ^ create_watch(Platform::String ^ strDirectory);
 
@@ -183,7 +183,9 @@ namespace file_watcher
 		m_watchmap.set_at(watchid, pwatch);
 
 		return watchid;
+
 	}
+
 
 	void os_file_watcher::remove_watch(const char * directory)
 	{
